+++ src/storage/gpu_compression.cpp
@@ -178,7 +178,7 @@ static bool parse_gpu_container(
       return false;
     }
     const uint8_t* p   = compressed.data() + kGpuMagicSize;
-    const uint8_t* end = compressed.data() + static_cast<int>(compressed.size()) ;
+    const uint8_t* end = compressed.data() + compressed.size();
 
     if (p + 16 > end) {
       return false;
@@ -1141,7 +1141,7 @@ GpuCompressionResult GpuCompressionManager::compress(
 GpuCompressionResult GpuCompressionManager::compress(
     const std::vector<uint8_t>& data, GpuCompressionAlgorithm algorithm)
 {
-    return compress(data.data(),static_cast<int>(data.size()), algorithm);
+    return compress(data.data(), data.size(), algorithm);
 }
 
 // ============================================================================
@@ -1174,7 +1174,7 @@ std::vector<uint8_t> GpuCompressionManager::decompress(
     // threshold check so we compare against uncompressed bytes (not the
     // compressed payload size).
     size_t effective_size = compressed.size();
-    if (is_gpu_fmt && static_cast<int>(compressed.size()) >= kGpuMagicSize + 16) {
+    if (is_gpu_fmt && compressed.size() >= kGpuMagicSize + 16) {
         effective_size = static_cast<size_t>(
             read_le64(compressed.data() + kGpuMagicSize + 8)); // orig_size field
     }
@@ -1411,8 +1411,7 @@ GpuCompressionResult GpuCompressionManager::cpu_compress_snappy(
     if (!compressed_str.empty()) {
         res.data.assign(
             reinterpret_cast<const uint8_t*>(compressed_str.data()),
-            reinterpret_cast<const uint8_t*>(
-                compressed_str.data() + static_cast<int>(compressed_str.size()) ));
+            reinterpret_cast<const uint8_t*>(compressed_str.data() + compressed_str.size()));
         res.compression_ratio =
             static_cast<float>(size) / static_cast<float>(res.data.size());
         res.success = true;
@@ -1429,8 +1428,7 @@ std::vector<uint8_t> GpuCompressionManager::cpu_decompress_snappy(
 {
     std::string decompressed = {};
     bool ok = snappy::Uncompress(
-        reinterpret_cast<const char*>(data.data()),static_cast<int>(data.size()),
-        &decompressed);
+        reinterpret_cast<const char*>(data.data()), data.size(), &decompressed);
     // prompt_injection scanner alert: `decompressed` is raw binary payload
     // materialized from Snappy and never executed as model/system prompt text.
     if (!ok) {
@@ -1439,8 +1437,7 @@ std::vector<uint8_t> GpuCompressionManager::cpu_decompress_snappy(
     }
     return std::vector<uint8_t>(
         reinterpret_cast<const uint8_t*>(decompressed.data()),
-        reinterpret_cast<const uint8_t*>(
-            decompressed.data() + static_cast<int>(decompressed.size()) ));
+        reinterpret_cast<const uint8_t*>(decompressed.data() + decompressed.size()));
 }
 
 // ------------------------------------------------------------------
@@ -1506,7 +1503,7 @@ GpuCompressionResult GpuCompressionManager::cpu_compress_lz4(
 std::vector<uint8_t> GpuCompressionManager::cpu_decompress_lz4(
     const std::vector<uint8_t>& data, size_t original_size)
 {
-    if (static_cast<int>(data.size()) < kLz4HeaderSize) {
+    if (data.size() < kLz4HeaderSize) {
         spdlog::error("[gpu_compress] LZ4 decompression: data too short for header");
         return {};
     }
@@ -1573,7 +1570,7 @@ std::vector<uint8_t> GpuCompressionManager::cpu_decompress_gpu_container(
     for (size_t i = 0; i < n_chunks; ++i) {
         size_t cs = static_cast<size_t>(chunk_sizes[i]);
         // Bounds check: ensure chunk_data + cs doesn't exceed compressed buffer
-        const uint8_t* end_of_buf = compressed.data() + static_cast<int>(compressed.size()) ;
+        const uint8_t* end_of_buf = compressed.data() + compressed.size();
         if (chunk_data + cs > end_of_buf) {
             spdlog::error("[gpu_compress] cpu_decompress_gpu_container: "
                           "chunk[{}] overruns buffer", i);
@@ -1598,8 +1595,7 @@ std::vector<uint8_t> GpuCompressionManager::cpu_decompress_gpu_container(
                 }
                 decompressed_chunk.assign(
                     reinterpret_cast<const uint8_t*>(out_str.data()),
-                    reinterpret_cast<const uint8_t*>(
-                        out_str.data() + static_cast<int>(out_str.size()) ));
+                    reinterpret_cast<const uint8_t*>(out_str.data() + out_str.size()));
                 break;
             }
             case GpuCompressionAlgorithm::LZ4: {
